+++ NEWS.md
@@ -1,6 +1,8 @@
 ### acefile 0.6.15-dev
 
 -   Add twine to dev requirements.
+-   Decode archive member filenames as CP850 by default and allow callers to
+    select a different filename encoding.
 
 
 ### acefile 0.6.14 2026-01-21